rename: storage 서비스들을 하나의 폴더로 묶음
